+++ App.css
@@ -1,8 +1,20 @@
-#root {
-  max-width: 1280px;
-  margin: 0 auto;
+html, body, #root {
+  height: 100%;
+  margin: 0;
+}
+
+.layout {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+}
+
+.mainContent {
+  flex: 1;
+  display: flex;
+  justify-content: center;
+  align-items: center;
   padding: 2rem;
-  text-align: center;
 }
 
 .card {
@@ -21,11 +33,9 @@
 }
 
 .centeredSection {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  min-height: calc(100vh - 80px);
-  padding: 2rem;
   text-align: center;
+}
+
+h1, h2 {
+  font-family: 'CatPaw', sans-serif;
 }
