+++ src/styles/HomePage.module.css
@@ -1,0 +1,6 @@
+.instruction {
+    margin: 1rem 0;
+    font-size: 1.1rem;
+    color: #444;
+}
+
